Improve 2018/yang/try.sh + fix make clobber
The script now compiles generated1.c and runs it. This might actually be
one of the others (some of the generated*.c files have the same effect)
but it's perhaps interesting to show nonetheless. This is only done if
the spoilers option is selected.

Update .gitignore with better globs.

Fixed make clobber to remove generated* rather than generated*.c.