shellcheck fixes in 2020/kurdyukov2
try.sh now runs makegif.sh.

Fix makegif.sh to first compile the program, allow one to specify CC and
convert and check that the output file was created. In the case
convert(1) is not found or it fails to generate the output file it
suggests the user look at the FAQ entry about ImageMagick.