Fix to use/try in 2011/richards/README.md
The alt try script is not mentioned in the README.md file as the alt
version is only a starting point for others who might wish to try and
fix this entry (as alternate code) for Macs with the Apple chip. The
alternate code I added as a starting point but the bugs.md is where the
details are. Since the alt code does not actually function properly it
shouldn't be referenced in the README.md file. It should only be
referenced in the bugs.md which it is (the bugs.md was slightly updated
with this too).