Improve 2005/giljade test.sh and fix make clobber
The script now lets one keep the out and out.count file even if the test
suite passes. This is done because the try script does that too in case
the user wants to see the out file to open in vim for fun (though the
try script actually simulates this for them).

The make clobber rule needs to remove out.count too in case they decide
to keep the file OR in the case before this commit the detecting errors
is terminated prematurely. This is especially important for updating the
manifest but even then it shouldn't leave files lying about.

Updated .gitignore with out.count added.