Update less options in 2001/ollinger/try.sh
The question is if I updated the less options by adding an option to
less(1) is it really less options? Perhaps it is more (or if you prefer:
more(1)) options for less? But since less is more, as they say, is it
really more options for more(1) or less options for more(1)? This leaves
more questions unanswered than answered questions (or questions
answered), more or less! :-)