Remove ambiguity in 2005/jetro/README.md
The wording 'this entry requires SDL to be installed' has more than one
interpretation: 'it requires SDL to install the entry' (which seems more
likely even though it's incorrect) or 'it requires SDL to be built'
(less likely as that is not what it said). Now it just says 'this entry
requires SDL'. Now it's monosemic and it says exactly what it should
(okay - it could say something else like 'it requires SDL to compile'
but that's bordering on unnecessary semantics).