change order of YYYY level index.html sections
The list of winners for a given year in that year's
top level `index.html` file will be the equivalent of the
file inventory of a winner's `index.html`.

A year's top level `index.html` "_Inventory_" link will
jump down to the lower section of the `index.html` file.
Thus, the output of `bin/output-year-index.sh` needs to come
after, this we need to use `-a` (not `-b`).