Minor fixes in 205fa25
It appears that indeed winners_text was meant to signify 'winner's text'
so 'entries_text' was changed to 'entry_text'. Also WINNER_TEXT variable
was changed to ENTRY_TEXT. If it is desired it can be changed to
'ENTRYS' but this is very wrong in my head and I can't get myself to do
it without it being declared significant enough to change.