Fixes and improvements in 2011/vik/try.sh
Check for perl(1) before trying to run ./raw2wav. If neither mplayer(1)
nor perl(1) are installed then the script tells the user to install one
or both and try again.

Function raw2wav() checks that perl(1) is installed. If it's not it
silently returns. With this change and the above it no longer makes any
assumption about what is installed and also tells the user what they
need to do if they don't have the proper tools installed.

Typo fix.