Add 1995/leo/try.sh and update bugs.md
There appears to be a bug where the program blocks. In one case another
command makes it work but the judges remarks suggest one way and it does
not. The other command does not seem to work at all, even if used like
the workaround with the first one.

Updated make clobber to remove foo.ps.

Updated .gitignore.