Finish editing of FAQ 1.2
This improves the names of links and adds some new links as well. A
point of interest that could easily be missed with a diff is that the
link of Yusuke Endoh was changed from his GitHub account to his
authors.html section. Also, since the first successful pull request to
the winner repo was acknowledged by author (Yusuke Endoh) I added the
author of the first successful pull request to the temp-test-ioccc repo
(as the commit was cited) (and that indeed would be me). It might be
that keeping the GitHub accounts would be better and I can certainly
change them back but it seemed nicer to acknowledge the author entries
as that's the real honour of being a winner.

There were other things besides the above but they were all along the
same lines.