Make 2004/burley fix more like original
In particular although main() has to have two args don't do (void)argc
as that was not in the code.

To make it more consistent with other added functions (that I had to do
in other entries for instance and for the same reason) I made the
function poke() (r short of poker) pain() instead.

The thanks file was modified for this entry too, improving it a bit
(hopefully making certain points clearer too).