Fix 1993/lmfjyh/Makefile
The problem it had is that it was reverted back to trying to compile the
very unsafe file name without deleting it after the ${CC} line. Thus the ${RM}
part was moved back to the end of the rule. I added the file (or a glob
that will match the file) to .gitignore a long time ago possibly when I
added the alt code.

The thanks file was fixed to account for the fact that the Makefile no
longer automatically builds the alt version if the code fails to
compile.