add .tar.tstamp to all entries
We added a `.tar.tstamp` file to all entries which will
hold the timestamp for use when forming entry compressed tarballs.

These entry `.tar.tstamp` files will be part of a fix to
build compressed tarballs that both macOS and Linux agree on.

We wanted to push out this update now so that it won't
conflict future changes to the IOCCC manifest.