Manifest updates and fixes
Change 'download entry' to 'download entry tarball' as 'download entry' can be
slightly ambiguous: is it the source file or is it the entire entry? Perhaps
'download entire entry' could also be used but that might be going too far.

Change 'markdown source for index.html' to 'markdown source for this web page'
to match the index.html description in the manifest and because it makes more
sense this way.

Changed order of a.out.h and gentab.c in 1984/mullender in manifest: since
a.out.h description refers to gentab.c (as it's used in gentab.c) list gentab.c
first.

With this commit notwithstanding any possible manifest changes (which would
likely be for all other entries too) the index.html files of the entries from
1984 should be considered completely reviewed. The 1984/index.html file might
be done too but this depends on a discussion taking place.