replace obsolete a name tags with div
This HTML is now obsolete in HTML 5:

```html
<a name="faq0"></a>Section 0: Submitting entries to a new IOCCC
```

Instead we need to use the div HTML elements:

```html
<div id="faq0">Section 0: Submitting entries to a new IOCCC</div>
```

We are testing a fix for the top level HTML files.

Run `make gen_top_html` to perform the above and test under macOS.

TODO: Apply this type of fix across the entire web site if all is well.