fix h2 div containment conflict
It turns out that one cannot have a div element inside an h2 element.

This does not work:

```html
<h2><div id="copyright">Copyright and CC BY-SA 4.0 License</div></h2>
```

Instead we need to do:

```html
<div id="copyright"><h2>Copyright and CC BY-SA 4.0 License</h2></div>
```

We are testing a fix for the top level HTML files.

Run `make gen_top_html` to perform the above and test under macOS.

TODO: Apply this type of fix across the entire web site if all is well.