Improve make help and typo fixes in Makefile
The make help rule now has a summary of each rule so that one need not
check the Makefile every time. Of course for more details they still
have to but it's helpful for those who know what they need to do but
still might not remember all the rule names (as there are quite a few).

Typo fixes include things like incorrect statements (in comments) of
what a rule does, some being yank paste errors.