!
Or more specifically: add missing sentence ending which happens to be a
'!'. Changed 'PLEASE don't' to 'PLEASE DO NOT' in bold (with NOT in
italics too) to stress the point that the judges do NOT want people to
program in the style of IOCCC programs (but where's the fun in that,
Landon? :-) okay I'm joking .. maybe :-) ).