Fix links/text in 2020/ferguson[12]
Some of the removed text referred to things that no longer exist due to
shortening the text (a bit).

Others were that the text used to say that one should try the markdown
or html file depending on if read in GitHub or not but now it's only the
html files so that reference to GitHub was removed.

Other problems were discovered and fixed as well.