Improve bin/gen-top-html.sh
It is now possible to specify select files to generate like the comment
at the top suggested (but was not actually allowed). This is useful
rarely but there is a case for it and that has been updated and it runs
a fair bit quicker now.

It is an error if no files specified are in the allowed set (the
variable TOP_MD_SET). If at least one file specified is in TOP_MD_SET it
will process that or those files but it will skip any that are not. This
is to prevent trying to process other markdown files like README.md
files without the proper script.