Don't build alt automatically in 1985/sicherman
This was done at first to try and protect against clang but this was
decided against. The thanks file has this updated too but it is not
being committed yet as there are many other fixes and more to be done
before that can happen. README.md was updated and index.html was
rebuilt, however.