Make fixes in thanks file through 1987
This does not mean I have checked every link. Most have been checked
but more checks will follow.

Also as for the fgets() changes in the file it links to the FAQ about
this. It felt like a good idea as this is one where it can seem like
(and a judgement call had to be made) tampering with the entry. This
felt a better way than having a section about the fgets() change in the
file which really isn't to do with thanks but rather why. With this
change one can look at a specific entry, see the change and then if they
have any concerns see the FAQ that's directly linked to right there.