Restore anonymous locations to anonymous authors
In the case of 1984, 2001 and 2005 the anonymous authors had location
codes but the country code for anonymous locations is XX. This has been
done and the authors.html, location.html and the respective index.html
files have been rebuilt.

One file was not touched because I was not sure how to best approach it
as an email was provided and possibly other things too (in particular a
second author of 2015/endoh1).