Update template issue forms
I believe that with this commit issue #858 can be marked complete but
this can only be determined once this has been merged. From a test repo
I made it does look pretty nice.

Things are more consistent, there is a place for attaching images
explicitly (though not required as it's not always relevant though very
helpful), typo fixes and rewording at the least.