Minor manifest update
Changed in 1984/anonymous 'source code tattoo image' to 'source code and
tattoo image'.

Changed in 1984/mullender 'header file for gentab.c for compatibility
with 1984 systems' to 'gentab.c header file for compatibility with
systems in 1984'. The rationale here is that it's not 1984 systems but
systems in 1984. I would have kept the ordering of the wording the same
but that was more than 59 characters. The way it is worded now is not
necessarily better (other than not being literally incorrect - that is
if it's interpreted literally it is incorrect) but it's also not worse.

Minor format fix in 1985/shapiro/README.md.

Rebuilt index.html and .entry.json files.