correctly process non-zero EXIT_CODE, fix 2020/endoh2/README.md
We found cases where pandoc reported a conversion from markdown to
HTML problem that, while reported by the `bin/pandoc-wrapper.sh`
to `bin/md2html.sh`, which in turn reported to error to the tool
calling it, the tool calling `bin/md2html.sh` set `$EXIT_CODE` to
a non-zero value in order to delay the abort but then failed to act
on the non-zero `$EXIT_CODE` value at the end.

In a number of cases when `$EXIT_CODE` was set to non-zero, the
stderr message was labeled as a "Warning" instead of an "ERROR".

Fixed a bug in `bin/gen-authors.sh` that resulted the generation
of invalid markdown/HTML.  An `<hr>` HTML tag inside a `<div>` span
caused pandoc to issue "Div at .. closing implicitly" at the _hr_
HTML tag line.  For some reason, one cannot place a `<hr>`
(horizontal line) inside a `<div ... /div>` span.

Fixed a broken markdown links in `2020/endoh2/README.md` whose error
was hidden by one of the above mentioned problems.

Changed `bin/gen-authors.sh` VERSION from "1.10.1 2024-05-07"
to "1.10.2 2024-05-19" as a result of the above changes.

Changed `bin/gen-location.sh` VERSION from "1.4.1 2024-04-28"
to "1.4.2 2024-05-19" as a result of the above changes.

Changed `bin/gen-top-html.sh` VERSION from "1.4.6 2024-05-05"
to "1.4.5 2024-05-19" as a result of the above changes.

Changed `bin/gen-years.sh` VERSION from "1.3.1 2024-04-28"
to "1.3.2 2024-05-19" as a result of the above changes.

Changed `bin/quick-readme2index.sh` VERSION from "1.1.4 2024-04-13"
to "1.1.5 2024-05-19" as a result of the above changes.

Changed `bin/readme2index.sh` VERSION from "1.2.4 2024-04-14"
to "1.2.5 2024-05-19" as a result of the above changes.

Performed `make www` to test the above under macOS.