Many fixes in thanks file through 1995
This includes broken links and typos but also a stylistic change as
well. In particular each index.html file is preceded with the text:

    Winning entry:

and the next line, the source code, is preceded with:

    Winning entry source code:

This is to make them stand out more. It's not, maybe, immediately
obvious that the first links to index.html but I have not come up with a
way to make it obvious that looks nice as adding 'index.html' to
'Winning entry' might take from the fact that the section below is about
the fixes to the entry (or I might be taking it too literally too which
is entirely possible :-) ).

Missing links were also added for the years beyond 1995 at least for the
YYYY/index.html links. Those have not been checked for typos but when I
get to those years I will check those too.

I hope to get this file done in another two edits but probably no more
than three will be required.

Regenerated thanks-for-help.html file.