improve "is for" headers in authors.html
The "letter is for" is now inside HTML italics and
the C keyword is now inside HTML code.

Using lower case letters as the C keywords are
also in lower case and we prefer lower case.

Changed `bin/gen-authors.sh` VERSION from "1.10.2 2024-05-19"
to "1.10.3 2024-05-24" to perform the above.

Perform `make gen_authors` to perform the above.