Minor fixes and improvements in thanks file
For each year it now shows which IOCCC it was. In other words for 1984
the link title is now:

    1984 - The 1st IOCCC

and 1985 is:

    1985 - The 2nd IOCCC

and 1986 is:

    1986 - The 3rd IOCCC

and so on up to 2020 which is:

    2020 - The 27th IOCCC

This helps it stand out better and it shows which contest it was.

There were a few other fixes. This edit does NOT check further years
from yesterday's commit 2647827d94a48d89cd70c5dfb387d32fb9127bec and it
is possible I won't do more years until another day (though I started
1996 but nothing needed to be fixed, I think).

Other minor fixes in formatting were also made.

Rebuilt thanks-for-help.html.