Improve 1996/schweikh2/try.sh
As I was going through another file I noticed that something was missing
from the script, if perl is installed, and it also prompts the user if
they wish to run the author's suggested command (it results in an
infinite loop which is why it was left out but prompting for it is okay
as it's at the end).

The make clobber rule was updated for this update.