Improve FAQ 3.0 - about Makefile rules
It now clarifies that the CC=... checks by substring so that one can do
make CC=gcc-mp-12, for example, and it will register as gcc.

It also clarifies that the top level Makefile and the year Makefiles
recurse into subdirectories so that one can compile individual entries,
entries for a given year or all entries.