Fix some border cases of 'alt' to 'alternate'
I did not think of 'alt' at end of the line or 'alt version' (often if
not always at end of the line). This has been corrected.

As well I forgot to sort the manifest.numbers spreadsheet but that will
be in a commit that will be made in a bit.