Minor fixes in FAQ 3.20 + typo fixes
In the FAQ entry for downloading the tarballs (which has been renamed to
be a bit clearer) the example tarball for each kind (individual entry
and an entire year) it now links to the tarball in question so one can
more easily determine the path. It also shows where one can find the
tarballs (for each type). Other improvements have also been made but
this and the rest of the file will be further reviewed at another time.

Typo fixes.