Restore one INABIAF in 1990/westley
The other one cannot be restored or at least should not be as it floods
the screen in an unsuspecting way. If no arg is specified though, it now
will likely crash (previously because of the use of argv[1] later in the
code but now to prevent the code from flooding the screen if the number
is < 0).