Update formatting in next/README.md
This is not yet complete as I must still verify that link titles are
consistent with what they link to and formatting might also be changed a
bit (bold, italic) to help things stand out a bit better. An example of
a format change in this commit is that the 'status is foo' now has the
status (as in foo e.g. closed, open etc.) is in bold and italic (in the
header it is just italic as there seems to be no reason to be bold
there).

Rebuilt next/index.html.