Fix typos in next/README.md, run make gen_top_html
One typo was fixed and wording was slightly changed in one thing which
is not really a typo but might be clearer now (if only slightly).

The links seem okay though I did not go through them in the browser as
each section has the same links and it is easier to check them in vim
than going back and forth in the browser between pages. It should all be
good.

Thus next/README.md and next/index.html have been updated and can be
considered completely reviewed at this time.