Check links/formatting in remaining next/ files
Some links were fixed. For instance instead of index.html#enter it's now
the relevant FAQ. A markdown link was broken so that it did not create a
link at all.

Some formatting was updated mostly to make points bold or in code block
(if it's going to be html it might as well especially as they are
important).

And although there were some typo fixes there still are some typos in
there. If you don't know why that is see the FAQ 1.0: How did the IOCCC
get started?. :-) Besides much will be changed in these files by the
judges and the purpose of this commit was to make sure links were valid
and that the files are rendered okay, not to try and change the rules
(although I did update the IOCCC email and make it the 28th IOCCC and
fix the judges [list] .. to make a reference to something recent :-) ).

This commit means that the files under next/ have all been checked. As
usual we all anticipate things to change but some things that were
outdated were corrected (or they were at least partially corrected like
referring to the mkiocccentry repo) to help out a bit.