Mostly finish bin/README.md
Usage for scripts where the usage was missing added, even if that's just
that it is used in another file.

The script links have been changed to use %%REPO_URL%% for the GitHub
syntax highlighting. Some other files have also been changed to this.

Other changes were probably made too.

Links still have to be checked but this has to be done next time.