Fix links/formatting in markdown.md
Something that was not noted in the other commit that touched this file
is that because this is markdown which renders to html I changed the
comments in the form of:

	<=== no thank you

to

	<!-- no thank you -->

This was not strictly necessary but and for non html it might not even
make sense except that it is still rendered to html and in most cases
(if not all) it is html. This was done for the rest of the file now too.