Fix two authors' info in 1995, update FAQ
The (same two) authors of 1995/heathbar and 1995/makarios had invalid
information, in one case email as well as in both the url and location.

The location was previously set to unknown (ZZ) but I knew from the
archive (which used to be what the README.md files which were the hint
files had until I was asked to remove that information which was done
November 2023) that it was US.

I also noticed, however, that the url was invalid and in one case the
email was invalid (it referred to the same domain).

With these problems in mind the url was changed to refer to the website
on the Internet Wayback Machine (in both), the email was set to null in
the one that had an email (now it's null in both) and the location_code
was set to US (in both).

I then did:

	touch 1995/heathbar/README.md 1995/makarios/README.md

so that I could do:

	make quick_www

as only two index.html files had to be updated. This also updated
authors.html and location.html, of course.

The FAQ was updated with more information on how to update author
information with the above two as an example and it now also says how to
validate json for those who want to make changes to author information
(or if anything else in the future is put in json).