Fix author URLs
All the authors now have either a null URL, a valid URL or one on the
Internet Wayback Machine.

Where it seemed likely (say because the domain no longer exists) emails
were set to null also.

Now unless a bug is found in the script this really should finish the
checks of authors.html.