Improvements in formatting of guidelines
There were several places where a code block was used for text that is
more like a quote (and in one case it is a quote) and these have been
converted to a blockquote.

A few links were added instead of having the link in a code block.

In a case where a paragraph was split with blank lines in between each
line (with the '|' marker at each line) I added a <br> to it so that it
looked more natural.

With the recent changes to this file and the rules file more checks have
to be made as well.