improve IOCCC rules and guidelines
Fixed a number of typos, inconsistencies, and problems with the
rules and guidelines.

Improved how the closed version of the rules and guidelines declare
themselves at the top of the markdown and HTML file.

Changed `next/rules.md` VERSION from "2024-06-29-v28"
to "28.1 2024-06-29" in order to be more consistent
with IOCCC tool version format.

Changed `next/guidelines.md` VERSION from "2024-06-29-v28"
to "28.1 2024-06-29" in order to be more consistent
with IOCCC tool version format.

Make a minor format fix to `faq.md`.

Performed `make status` and `make www` to test the above umder macOS.