Fix technical details with rules (esp 17)
There were numerous things that had to be fixed and there will likely
have to be some discussion as well.

Some formatting fixes and additional links were also added.

This commit involves some updates to the FAQ and while I was at it I
fixed some references of 'entry'/'entries' to 'submission'/'submissions'
(those that should not be 'entry' or 'entries').