Update rules wrt mastodon
Unless one is mentioned one will not get a push notification so one
should refresh the IOCCC mastodon feed every so often. This is at the
top of the file, though, and seemingly part of a file in inc/ so this
might have to be done differently than how this change was done.