update rules and guideline headers
Harmonized and improve the headers that are prepended
to the rules and guidelines given the IOCCC state.

Attempted to make the headers that are prepended
to the rules and guidelines more consistent.

Make reference to, where applicable, FAQs
about feedback and questions in those headers.

Perform `make www` to perform the above and test under macOS.