Add issue template, fix others
The documentation template is for a missing documentation file or
problem with documentation. This MIGHT be classified under a website
issue but it is kind of a topic on its own and as an issue for this was
recently opened (with the documentation label) this seems like a worthy
thing to have.

The bug issue template now properly adds the 'bug' label.

The other issue templates have lower case labels.